Satchel concealment (#186)
## About The Pull Request
reimplements concealment of satchels.
alt-click to conceal the satchel. Must be on your back, and must be
wearing a uniform.
holding an item in your hand, having the backpack anywhere except for
your back, or not having a uniform equipped will result in an alt-click
opening the storage as normal.
## Why It's Good For The Game
the backpacks are ugly on most outfits (especially suits) so many
players use satchels and conceal them on the old codebases where this is
a feature.
## Changelog
:cl:
add: alt-click to conceal/reveal satchels
/:cl: